refactor(omegaform): extract walker into meta/walker.ts
Moves the five walker-private/walker-public definitions (isNullishType,
unwrapSingleLiteralUnion, unwrapNestedUnions, WalkerContext, ParentMeta,
leafMetaForAst, walkStruct, classifyAndWalkUnion, walk) out of
createMeta.ts into a new sibling file meta/walker.ts, bringing
createMeta.ts from 435 lines down to 185.
